Wordle Solver
Fun little exercise to generate high value guesses for the first turn on
Wordle

This would benefit from popularity information from words, such as
frequency of words appearing in corpora of US English texts, as a
filtering and ranking input. Without that popularity information, the
words here are real words but are kind of Scrabble words. Will have to
try guessing some of these words and seeing how it goes. Also
considering other words to add more scores to each individual word and
scoring the word.